[fix] 데이터 테이블 수정

+++ Assets/Scripts/Objects/Monster/Monster.cs
@@ -9,7 +9,7 @@ public class Monster : MonoBehaviour
     // TODO need refactor this 
     [SerializeField] private int dataKey;
     public List<SpriteRenderer> spriteRenderers { get; private set; }
-    public int CurrentHp { get; set; }
+    public float CurrentHp { get; set; }
     [HideInInspector] public MonsterTable monsterData;
 
     public Animator Animator;
